+++ application/edge_agent/components/app_config/include/app_config.h
@@ -28,6 +28,7 @@ typedef struct {
     char llm_base_url[APP_CONFIG_STR_LEN];
     char llm_auth_type[32];
     char llm_timeout_ms[16];
+    char llm_max_tokens[16];
     char qq_app_id[32];
     char qq_app_secret[APP_CONFIG_STR_LEN];
     char feishu_app_id[64];