+++ components/claw_modules/claw_core/src/claw_core_llm.c
@@ -61,6 +61,7 @@ esp_err_t claw_core_llm_init(const claw_core_llm_config_t *config, char **out_er
     runtime_config.base_url = config->base_url;
     runtime_config.auth_type = config->auth_type;
     runtime_config.timeout_ms = config->timeout_ms;
+    runtime_config.max_tokens = config->max_tokens;
     runtime_config.image_max_bytes = config->image_max_bytes;
     return claw_llm_runtime_init(&s_runtime, &runtime_config, out_error_message);
 }