+++ components/common/app_claw/include/app_claw.h
@@ -29,6 +29,7 @@ typedef struct {
     char llm_base_url[APP_CLAW_STR_LEN];
     char llm_auth_type[APP_CLAW_SHORT_STR_LEN];
     char llm_timeout_ms[APP_CLAW_TIMEOUT_LEN];
+    char llm_max_tokens[APP_CLAW_TIMEOUT_LEN];
     char qq_app_id[APP_CLAW_SHORT_STR_LEN];
     char qq_app_secret[APP_CLAW_STR_LEN];
     char feishu_app_id[APP_CLAW_MODEL_LEN];